+++ cookbooks/utils/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file is used to list changes made in each version of utils.
 
+## 0.2.0:
+
+* recipe/hardinfo - manages hardinfo app installation
+* recipe/p7zip - manages p7zip-full app installation
+
 ## 0.1.0:
 
 * Initial release of utils